refactor: consolidate weighted_f1_score func into lm_eval utils
