+++ lm_eval/tasks/afrimmlu/utils.py
@@ -1,3 +1,7 @@
+from lm_eval.utils import weighted_f1_score
+
+
+
 def doc_to_choice(doc):
     choices = eval(doc["choices"])
     return choices
@@ -27,13 +31,3 @@ def doc_to_text(doc):
         choice4=choices[3],
     )
     return text
-
-
-def weighted_f1_score(items):
-    from sklearn.metrics import f1_score
-
-    unzipped_list = list(zip(*items))
-    golds = unzipped_list[0]
-    preds = unzipped_list[1]
-    fscore = f1_score(golds, preds, average="weighted")
-    return fscore
