+++ lm_eval/tasks/afrixnli/anli prompt/native-direct/utils.py
@@ -1,8 +1,2 @@
-def weighted_f1_score(items):
-    from sklearn.metrics import f1_score
+from lm_eval.utils import weighted_f1_score
 
-    unzipped_list = list(zip(*items))
-    golds = unzipped_list[0]
-    preds = unzipped_list[1]
-    fscore = f1_score(golds, preds, average="weighted")
-    return fscore
