+++ lm_eval/tasks/afrixnli/lai prompt/direct/utils.py
@@ -1,3 +1,7 @@
+from lm_eval.utils import weighted_f1_score
+
+
+
 def doc_to_text(doc):
     output = """Please identify whether the premise entails or contradicts the hypothesis in the following premise
     and hypothesis. The answer should be exact entailment, contradiction, or neutral.
@@ -15,12 +19,3 @@ def doc_to_target(doc):
     replacements = {0: "entailment", 1: "neutral", 2: "contradiction"}
     return replacements[doc["label"]]
 
-
-def weighted_f1_score(items):
-    from sklearn.metrics import f1_score
-
-    unzipped_list = list(zip(*items))
-    golds = unzipped_list[0]
-    preds = unzipped_list[1]
-    fscore = f1_score(golds, preds, average="weighted")
-    return fscore
